+++ src/instamatic/calibrate/calibrate_beamshift.py
@@ -342,9 +342,9 @@ def main_entry():
     args = options.args
 
     if not args:
-        from instamatic import TEMController
+        from instamatic import controller
 
-        ctrl = TEMController.initialize()
+        ctrl = controller.initialize()
         calibrate_beamshift(ctrl=ctrl, save_images=True)
     else:
         center_fn = args[0]
