+++ src/instamatic/calibrate/calibrate_brightness.py
@@ -206,9 +206,9 @@ def main_entry():
     args = options.args
 
     if not args:
-        from instamatic import TEMController
+        from instamatic import controller
 
-        ctrl = TEMController.initialize()
+        ctrl = controller.initialize()
         calibrate_brightness(ctrl, save_images=True)
     else:
         calibrate_brightness(args)
