+++ src/instamatic/calibrate/calibrate_stage_lowmag.py
@@ -430,9 +430,9 @@ def main_entry():
     args = options.args
 
     if not args:
-        from instamatic import TEMController
+        from instamatic import controller
 
-        ctrl = TEMController.initialize()
+        ctrl = TEMtroller.initialize()
         calibrate_stage_lowmag(ctrl=ctrl, save_images=True)
     else:
         center_fn = args[0]
