+++ src/instamatic/calibrate/calibrate_stage_mag1.py
@@ -289,9 +289,9 @@ def main_entry():
     args = options.args
 
     if not args:
-        from instamatic import TEMController
+        from instamatic import controller
 
-        ctrl = TEMController.initialize()
+        ctrl = controller.initialize()
         calibrate_mag1(ctrl=ctrl, save_images=True)
     else:
         center_fn = args[0]
