+++ src/instamatic/config/autoconfig.py
@@ -126,8 +126,8 @@ It establishes a connection to the microscope and reads out the camera lengths a
         cam_name = None
 
     from instamatic.camera.camera import get_cam
-    from instamatic.TEMController.microscope import get_tem
-    from instamatic.TEMController.TEMController import TEMController
+    from instamatic.microscope.microscope import get_tem
+    from instamatic.controller import TEMController
 
     if cam_connect:
         cam = get_cam(cam_name)() if cam_name else None
