+++ src/instamatic/controller.py
@@ -13,13 +13,13 @@ from instamatic.camera.camera_base import CameraBase
 from instamatic.exceptions import TEMControllerError
 from instamatic.formats import write_tiff
 from instamatic.image_utils import rotate_image
-from instamatic.TEMController.microscope_base import MicroscopeBase
+from instamatic.microscope.base import MicroscopeBase
 
-from .deflectors import *
-from .lenses import *
-from .microscope import Microscope
-from .stage import *
-from .states import *
+from instamatic.microscope.components.deflectors import *
+from instamatic.microscope.components.lenses import *
+from instamatic.microscope.microscope import Microscope
+from instamatic.microscope.components.stage import *
+from instamatic.microscope.components.states import *
 
 _ctrl = None  # store reference of ctrl so it can be accessed without re-initializing
