+++ src/instamatic/gui/ctrl_frame.py
@@ -187,9 +187,9 @@ class ExperimentalCtrl(LabelFrame):
 
         frame.pack(side='top', fill='x', padx=10, pady=10)
 
-        from instamatic import TEMController
+        from instamatic import controller
 
-        self.ctrl = TEMController.get_instance()
+        self.ctrl = controller.get_instance()
 
     def init_vars(self):
         self.var_negative_angle = DoubleVar(value=-40)
