+++ src/instamatic/microscope/__init__.py
@@ -1,4 +1,3 @@
 from __future__ import annotations
 
 from .microscope import Microscope
-from .TEMController import get_instance, initialize
