+++ src/instamatic/microscope/base.py
