+++ src/instamatic/microscope/components/deflectors.py
