+++ src/instamatic/microscope/components/lenses.py
