+++ src/instamatic/microscope/components/stage.py
