+++ src/instamatic/microscope/components/states.py
