+++ src/instamatic/microscope/interface/fei_microscope.py
@@ -9,7 +9,7 @@ from numpy import pi
 
 from instamatic import config
 from instamatic.exceptions import FEIValueError, TEMCommunicationError
-from instamatic.TEMController.microscope_base import MicroscopeBase
+from instamatic.microscope.base import MicroscopeBase
 
 logger = logging.getLogger(__name__)
