+++ src/instamatic/server/TEMServer_FEI.py
@@ -10,7 +10,7 @@ import subprocess as sp
 import threading
 from socket import *
 
-from instamatic import TEMController, config
+from instamatic import TEMtroller, config
 
 HOST = config.settings.fei_server_host
 PORT = config.settings.fei_server_port
@@ -68,7 +68,7 @@ Utility script to enable rotation control from a dmscript. See [https://github.c
     logging.captureWarnings(True)
     log = logging.getLogger(__name__)
 
-    ctrl = TEMController.initialize()
+    ctrl = TEMtroller.initialize()
 
     s = socket(AF_INET, SOCK_STREAM)
     s.bind((HOST, PORT))
