+++ tests/conftest.py
@@ -8,7 +8,7 @@ import pytest
 
 # Actually importing the controller class breaks the tests
 if TYPE_CHECKING:
-    from instamatic.TEMController.TEMController import TEMController
+    from instamatic.controller import TEMController
 
 
 base_drc = Path(__file__).parent
@@ -21,7 +21,7 @@ def pytest_configure():
 
 @pytest.fixture(scope='module')
 def ctrl() -> 'TEMController':
-    from instamatic.TEMController import initialize
+    from instamatic.controller import initialize
 
     ctrl = initialize()
